+++ scripts/pause.gd
@@ -24,4 +24,5 @@ func exit():
 	unpause()
 	var menu = load("res://scenes/debug_launcher.tscn").instantiate()
 	get_tree().root.add_child(menu)
-	get_parent().queue_free()
+	get_node("/root/TitleScreen").resumeMusic()
+	get_node("/root/Level").queue_free()