+++ scripts/screens/title.gd
@@ -1,8 +1,13 @@
 extends Node2D
 
+@onready var settingsScreen = load("res://scenes/screens/settings.tscn").instantiate()
+
 func _ready():
 	$TitleMusic.play()
 	$Version.text = "V"+ ProjectSettings.get_setting("application/config/version")
+	
+	settingsScreen.visible = false
+	get_tree().root.add_child.call_deferred(settingsScreen)
 
 func _on_arena_button_pressed():
 	var arenaSetupScreen = load("res://scenes/debug_launcher.tscn").instantiate()
@@ -24,3 +29,8 @@ func resumeMusic():
 func restart():
 	visible = true
 	$TitleMusic.play(0)
+
+
+func _on_settings_button_pressed():
+	settingsScreen.visible = true
+	visible = false